[refact] Api호출 및 폰트에 대한 1차 리팩토링
