+++ front/pages/Setting/SettingsView.js
