+++ front/pages/WriteDiary/WriteDiaryView.js
@@ -1,14 +1,10 @@
-import AppLoading from 'expo-app-loading';
-import { useFonts } from 'expo-font';
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import { View, Text, StyleSheet, Image, Dimensions, TextInput, Button, Keyboard } from 'react-native';
 import { TouchableOpacity, TouchableWithoutFeedback } from 'react-native-gesture-handler';
 
-import axios from 'axios';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { basic_theme } from '../theme';
-const baseUrl = 'http://152.67.193.252';
-const diaryUrl = '/diary/write';
+import { basic_theme } from '../../theme';
+import { axios_post } from '../../api/api';
 
 const WriteDiaryView = ({ navigation, date }) => {
   const [userId, setUserId] = useState('');
@@ -28,23 +24,7 @@ const WriteDiaryView = ({ navigation, date }) => {
       alert('내용을 입력해주세요');
     }
     // alert(`${userId}, ${date}, ${weather}, ${title}, ${contents}`); //확인용
-
-    const response = await axios.post(
-      `${baseUrl}${diaryUrl}`,
-      {
-        // 서버통신
-        userId: userId,
-        date: date,
-        weather: weather,
-        title: title,
-        contents: contents,
-      },
-      {
-        headers: {
-          'Content-Type': `application/json`,
-        },
-      }
-    );
+    const response = axios_post('write', { userId, date, weather, title, contents });
     if (response.status == 201) {
       navigation.navigate('AnalysisLoadingView', {
         diaryId: {
@@ -53,14 +33,6 @@ const WriteDiaryView = ({ navigation, date }) => {
       });
     }
   };
-  let [fontsLoaded] = useFonts({
-    //폰트 가져오기
-    Gowun_Batang: require('../assets/fonts/GowunBatang-Regular.ttf'),
-  });
-  if (!fontsLoaded) {
-    //폰트 가져오는 동안 AppLoading (local이라 짧은시간)
-    return <AppLoading />;
-  }
 
   console.log(contents);
